fix slider
